\name{evilDice-package}
\alias{evilDice-package}
\docType{package}
\title{
\packageTitle{evilDice}
}
\description{
\packageDescription{evilDice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{evilDice}
\packageIndices{evilDice}
Intransitive dice are dice which display the unintuitive property that every die has a winning probability against another die and a losing probability against another die in the set.  This function tests the input set, returning the pairwise probabilities along with other interesting data. 
 
}
\author{
\packageAuthor{evilDice} 
Some fast algorithm code supplied by user "onyambu" at  \url{https://stackoverflow.com/questions/75463792/}

Maintainer: \packageMaintainer{evilDice}
}
\references{
\url{https://en.wikipedia.org/wiki/Intransitive_dice}
}
